fix: move sync.go tests from undo_test.go to sync_helpers_test.go, delete undo_test.go
